Use custom fail type pattern
