+++ grin.toml
@@ -75,7 +75,7 @@ port = 13414
 log_to_stdout = true
 
 # Log level for stdout: Critical, Error, Warning, Info, Debug, Trace
-stdout_log_level = "Warning"
+stdout_log_level = "Debug"
 
 # Whether to log to a file
 log_to_file = true
