+++ wallet/Cargo.toml
@@ -17,6 +17,8 @@ serde = "~1.0.8"
 serde_derive = "~1.0.8"
 serde_json = "~1.0.8"
 bodyparser = "~0.7.0"
+failure = "0.1.1"
+failure_derive = "0.1.1"
 futures = "^0.1.15"
 iron = "~0.5.1"
 hyper = "~0.11.4"
